Rename lib/babel.js to lib/test-worker.js
