+++ lib/fork.js
@@ -9,7 +9,7 @@ var send = require('./send');
 module.exports = function (file, opts) {
 	opts = objectAssign({file: file}, opts);
 
-	var ps = childProcess.fork(path.join(__dirname, 'babel.js'), [JSON.stringify(opts)], {
+	var ps = childProcess.fork(path.join(__dirname, 'test-worker.js'), [JSON.stringify(opts)], {
 		cwd: path.dirname(file),
 		stdio: ['ignore', process.stderr, process.stderr]
 	});
