+++ lib/test-worker.js
